feature/split-auth-form: разделил форму входа, добавил отслеживание по урлу
