+++ src/feature/auth/authenticationForm/index.ts
@@ -1,0 +1,1 @@
+export { AuthenticationForm } from './AuthenticationForm';
