+++ src/feature/auth/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useAuthActions } from './useAuthActions';
