+++ src/feature/auth/registrationForm/index.ts
@@ -1,0 +1,1 @@
+export { RegistrationForm } from './RegistrationForm';
