+++ src/feature/auth/types/auth.type.ts
@@ -1,0 +1,7 @@
+export const loginFormFieldEnum = {
+    EMAIL: 'email',
+    LOGIN: 'login',
+    PASSWORD: 'password',
+} as const;
+
+export type LoginFormFieldEnum = (typeof loginFormFieldEnum)[keyof typeof loginFormFieldEnum];
